refactor(node): extract wallet subcommands to commands/wallet.rs
main.rs had grown to 4.78k LOC with ~35 cmd_* functions inline
between clap parsing and the async runtime bootstrap. Splitting one
logical group per file makes each command independently
reviewable.

This PR moves the six `sentrix wallet *` subcommands (+ the two
password-resolution helpers they share with cmd_start) out of
main.rs into `bin/sentrix/src/commands/wallet.rs`:

  cmd_wallet_generate, cmd_wallet_import, cmd_wallet_info,
  cmd_wallet_encrypt, cmd_wallet_decrypt, cmd_wallet_rekey
  + resolve_password, resolve_password_named

`commands::wallet::resolve_password` is also called from
main.rs::cmd_start (validator boot keystore decrypt) — keeping the
helpers in one module so prompt UX + env-var names can't drift
between callers.

`get_wallets_dir` widened from `fn` to `pub(crate) fn` so the new
module can read the operator's configured wallets dir via the same
SENTRIX_DATA_DIR-aware path. No other visibility changes.

CLI dispatch in main.rs updated to call
`commands::wallet::cmd_wallet_*` instead of bare `cmd_wallet_*`.

main.rs: 4,780 → 4,552 LOC (−228 = −4.8%).

Same shape as the sentrix-core blockchain.rs decomposition PRs:
refactor-only, no behaviour change, no consensus path touched.

Verification:
  cargo check --workspace -D warnings   clean
  cargo clippy -p sentrix-node --tests -D warnings   clean
  cargo test -p sentrix-node --release   11 passed (unchanged)
  pre-PR leak grep   clean
